Refactor: use barrel exports for services
(To create a kind of internal API)
