+++ src/services/settings/index.ts
@@ -1,0 +1,5 @@
+export * from "./defaultSettings";
+export * from "./editSettings";
+export * from "./getSettings";
+
+export * from "./types";
