+++ src/services/updateNotification/index.ts
@@ -1,0 +1,2 @@
+export * from "./getUpdateNotification";
+export * from "./setUpdateNotification";
