+++ src/utils/index.ts
@@ -1,3 +1,6 @@
 export * from "./cn";
 export * from "./date/formatToday";
 export * from "./date/formatWeekNumber";
+export * from "./getFavicon";
+export * from "./dailyRandomNumber";
+export * from "./normalizeUrl";
